requirements_production.txt
# AI Translator v2.2.5 - Production Requirements
flask>=2.3.0
flask-sqlalchemy>=3.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
werkzeug>=2.3.0
torch>=2.0.0
faster-whisper>=0.9.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
psutil>=5.9.0
pynvml>=11.5.0
paramiko>=3.2.0
boto3>=1.28.0
requests>=2.31.0
